refactor gmail.py into a baseclase SendMail for more sensical logging messages and also on-principle
