+++ parsons/notifications/gmail.py
@@ -1,0 +1,87 @@
+import base64
+from apiclient import errors
+from googleapiclient.discovery import build
+from httplib2 import Http
+from oauth2client import file, client, tools
+from parsons.notifications.sendmail import SendMail
+
+SCOPES = 'https://www.googleapis.com/auth/gmail.send'
+
+
+class Gmail(SendMail):
+    """Create a Gmail object, for sending emails.
+
+    `Args:`
+        creds_path: str
+            The path to the credentials.json file.
+        token_path: str
+            The path to the token.json file.
+    """
+
+    def __init__(self, creds_path=None, token_path=None):
+
+        self.user_id = 'me'
+
+        if not creds_path:
+            raise ValueError("Invalid path to credentials.json.")
+
+        if not token_path:
+            raise ValueError("Invalid path to token.json.")
+
+        self.store = file.Storage(token_path)
+        self.creds = self.store.get()
+
+        # BUG-1
+        # http = httplib2shim.Http()
+
+        if not self.creds or self.creds.invalid:
+            flow = client.flow_from_clientsecrets(creds_path, SCOPES)
+            self.creds = tools.run_flow(flow, self.store)
+
+            # BUG-1
+            # self.creds = self.run_flow(flow, self.store, http=http)
+
+        self.service = build('gmail', 'v1', http=self.creds.authorize(Http()))
+
+        # BUG-1
+        # self.service = build('gmail', 'v1', http=self.creds.authorize(http))
+
+    def _prepare_message(self, message, message_type):
+        msg = {'raw': base64.urlsafe_b64encode(message.as_bytes()).decode()}
+
+        self.log.debug(msg)
+        self.log.info(f"Encoded {message_type} sucessfully created.")
+
+        return msg
+
+    def send_message(self, message, user_id=None):
+        """Send an email message.
+
+        `Args:`
+            message: dict
+                Message to be sent as a base64url encode object.
+                i.e. the objects created by the create_* instance methods
+            user_id: str
+                Optional; User's email address. Defaults to the special value
+                "me" which is used to indicate the authenticated user.
+        `Returns:`
+            dict
+                A Users.messages object see `https://developers.google.com/gmail/api/v1/reference/users/messages#resource.` # noqa
+                for more info.
+        """
+        self.log.info("Sending a message...")
+        if not user_id:
+            user_id = self.user_id
+        try:
+            message = (self.service.users().messages()
+                       .send(userId=user_id, body=message).execute())
+        except errors.HttpError:
+            self.log.exception(
+                'An error occurred: while attempting to send a message.')
+            raise
+        else:
+            self.log.debug(message)
+            self.log.info(
+                f"Message sent succesfully (Message Id: {message['id']})")
+
+            return message
