+++ parsons/notifications/sendmail.py
@@ -1,5 +1,4 @@
 # Adapted from Gmail API tutorial https://developers.google.com/gmail/api
-from apiclient import errors
 from email.mime.text import MIMEText
 from email.mime.multipart import MIMEMultipart
 from email.mime.image import MIMEImage
@@ -8,11 +7,7 @@ from email.mime.base import MIMEBase
 from email.mime.application import MIMEApplication
 from email.encoders import encode_base64
 from email.utils import parseaddr
-from googleapiclient.discovery import build
-from httplib2 import Http
-from oauth2client import file, client, tools
 from validate_email import validate_email
-import base64
 import io
 import logging
 import mimetypes
@@ -28,54 +23,21 @@ import os
 
 logger = logging.getLogger(__name__)
 
-SCOPES = 'https://www.googleapis.com/auth/gmail.send'
 
-
-class Gmail(object):
-    """Create a Gmail object, for sending emails.
-
-    `Args:`
-        creds_path: str
-            The path to the credentials.json file.
-        token_path: str
-            The path to the token.json file.
+class SendMail(object):
+    """SendMail base class for sending emails.
     """
 
-    def __init__(self, creds_path=None, token_path=None):
+    log = logger
 
-        self.user_id = 'me'
-
-        if not creds_path:
-            raise ValueError("Invalid path to credentials.json.")
-
-        if not token_path:
-            raise ValueError("Invalid path to token.json.")
-
-        self.store = file.Storage(token_path)
-        self.creds = self.store.get()
-
-        # BUG-1
-        # http = httplib2shim.Http()
-
-        if not self.creds or self.creds.invalid:
-            flow = client.flow_from_clientsecrets(creds_path, SCOPES)
-            self.creds = tools.run_flow(flow, self.store)
-
-            # BUG-1
-            # self.creds = self.run_flow(flow, self.store, http=http)
-
-        self.service = build('gmail', 'v1', http=self.creds.authorize(Http()))
-
-        # BUG-1
-        # self.service = build('gmail', 'v1', http=self.creds.authorize(http))
+    def send_message(self, message):
+        raise NotImplementedError("send_message is how to send the prepared message")
 
     def _prepare_message(self, message, message_type):
-        msg = {'raw': base64.urlsafe_b64encode(message.as_bytes()).decode()}
-
-        logger.debug(msg)
-        logger.info(f"Encoded {message_type} sucessfully created.")
-
-        return msg
+        """Convert a email.mime-type object into an object format suitable for your backend to send
+           Consider overriding this, if necessary for your mail-sending backend
+        """
+        return message
 
     def create_message_simple(self, sender, to, subject, message_text):
         """Create a text-only message for an email.
@@ -92,7 +54,7 @@ class Gmail(object):
         `Returns:`
             An object passable to send_message to send
         """
-        logger.info("Creating a simple message...")
+        self.log.info("Creating a simple message...")
 
         message = MIMEText(message_text)
         message['to'] = to
@@ -119,7 +81,7 @@ class Gmail(object):
         `Returns:`
             An object passable to send_message to send
         """
-        logger.info("Creating an html message...")
+        self.log.info("Creating an html message...")
 
         message = MIMEMultipart('alternative')
         message['subject'] = subject
@@ -151,7 +113,7 @@ class Gmail(object):
         `Returns:`
             An object passable to send_message to send
         """
-        logger.info("Creating a message with attachments...")
+        self.log.info("Creating a message with attachments...")
 
         message = MIMEMultipart('alternative')
         message['to'] = to
@@ -180,7 +142,7 @@ class Gmail(object):
                 fp.close()
 
             content_type, encoding = mimetypes.guess_type(filename)
-            logger.debug(
+            self.log.debug(
                 f"(File: {f}, Content-type: {content_type}, "
                 f"Encoding: {encoding})")
 
@@ -190,24 +152,24 @@ class Gmail(object):
             main_type, sub_type = content_type.split('/', 1)
 
             if main_type == 'text':
-                logger.info("Added a text file.")
+                self.log.info("Added a text file.")
                 msg = MIMEText(file_bytes, _subtype=sub_type, _charset='utf-8')
 
             elif main_type == 'image':
-                logger.info("Added an image file.")
+                self.log.info("Added an image file.")
                 msg = MIMEImage(file_bytes, _subtype=sub_type)
                 msg.add_header('Content-ID', f'<{filename}>')
 
             elif main_type == 'audio':
-                logger.info("Added an audio file.")
+                self.log.info("Added an audio file.")
                 msg = MIMEAudio(file_bytes, _subtype=sub_type)
 
             elif main_type == 'application':
-                logger.info("Added an application file.")
+                self.log.info("Added an application file.")
                 msg = MIMEApplication(file_bytes, _subtype=sub_type)
 
             else:
-                logger.info("Added an unknown-type file.")
+                self.log.info("Added an unknown-type file.")
                 msg = MIMEBase(main_type, sub_type)
                 msg.set_payload(file_bytes)
                 encode_base64(msg)
@@ -219,7 +181,7 @@ class Gmail(object):
         return self._prepare_message(message, 'message with attachments')
 
     def _validate_email_string(self, str):
-        logger.debug(f"Validating email {str}...")
+        self.log.debug(f"Validating email {str}...")
         realname, email_addr = parseaddr(str)
 
         if not email_addr:
@@ -229,38 +191,6 @@ class Gmail(object):
             raise ValueError("Invalid email address.")
 
         return True
-
-    def send_message(self, message, user_id=None):
-        """Send an email message.
-
-        `Args:`
-            message: dict
-                Message to be sent as a base64url encode object.
-                i.e. the objects created by the create_* instance methods
-            user_id: str
-                Optional; User's email address. Defaults to the special value
-                "me" which is used to indicate the authenticated user.
-        `Returns:`
-            dict
-                A Users.messages object see `https://developers.google.com/gmail/api/v1/reference/users/messages#resource.` # noqa
-                for more info.
-        """
-        logger.info("Sending a message...")
-        if not user_id:
-            user_id = self.user_id
-        try:
-            message = (self.service.users().messages()
-                       .send(userId=user_id, body=message).execute())
-        except errors.HttpError:
-            logger.exception(
-                'An error occurred: while attempting to send a message.')
-            raise
-        else:
-            logger.debug(message)
-            logger.info(
-                f"Message sent succesfully (Message Id: {message['id']})")
-
-            return message
 
     def send_email(self, sender, to, subject, message_text, message_html=None,
                    files=None):
@@ -286,9 +216,9 @@ class Gmail(object):
         `Returns:`
             None
         """
-        logger.info("Preparing to send and email...")
+        self.log.info("Preparing to send and email...")
 
-        logger.info("Validating email(s)")
+        self.log.info("Validating email(s)")
         if isinstance(to, list):
             if len(to) == 0:
                 raise EmptyListError("Must contain at least 1 email.")
@@ -317,11 +247,11 @@ class Gmail(object):
             msg = self.create_message_attachments(
                 sender, to, subject, message_text, files, message_html)
 
-        logger.info(f"Sending a(n) {msg_type} email...")
+        self.log.info(f"Sending a(n) {msg_type} email...")
 
         self.send_message(msg)
 
-        logger.info("Email sent succesfully.")
+        self.log.info("Email sent succesfully.")
 
 
 class EmptyListError(IndexError):
