+++ parsons/notifications/smtp.py
@@ -1,12 +1,10 @@
 import smtplib
 
-from parsons.notifications.gmail import Gmail
+from parsons.notifications.sendmail import SendMail
 from parsons.utilities.check_env import check
 
-# There's some irony in this inheriting from Gmail, but life is strange sometimes
 
-
-class SMTP(Gmail):
+class SMTP(SendMail):
     """Create a SMTP object, for sending emails.
 
     `Args:`
